refactor(deps): move DisplayNames abstract operations to ecma402-abstract
- Move CanonicalCodeForDisplayNames and IsValidDateTimeFieldCode from
  intl-displaynames/abstract/ to ecma402-abstract/DisplayNames/
- Remove aggregate js_library — consumers depend on sub-packages directly
- Add explicit //packages/ecma402-abstract/types and /NumberFormat deps
  to all consumer packages that import from those sub-paths
- Delete intl-displaynames/abstract/ directory
