+++ packages/ecma402-abstract/tsconfig.json
@@ -5,14 +5,22 @@
       "#packages/*": ["../../packages/*"]
     }
   },
-  "exclude": ["./NumberFormat/**/*", "./scripts/**/*", "./types/**/*"],
+  "exclude": [
+    "./DisplayNames/**/*",
+    "./NumberFormat/**/*",
+    "./scripts/**/*",
+    "./types/**/*"
+  ],
   "extends": "../../tsconfig.json",
   "references": [
     {
-      "path": "./types/tsconfig.json"
+      "path": "./DisplayNames/tsconfig.json"
     },
     {
       "path": "./NumberFormat/tsconfig.json"
+    },
+    {
+      "path": "./types/tsconfig.json"
     }
   ]
 }
