+++ packages/icu-messageformat-parser/BUILD.bazel
@@ -21,6 +21,7 @@ SRCS = glob(["*.ts"])
 
 SRC_DEPS = [
     "//packages/ecma402-abstract",
+    "//packages/ecma402-abstract/types",
     "//:node_modules/@formatjs/icu-skeleton-parser",
 ]
