+++ packages/intl-displaynames/BUILD.bazel
@@ -36,6 +36,8 @@ TESTS = glob([
 
 SRC_DEPS = [
     "//packages/ecma402-abstract",
+    "//packages/ecma402-abstract/types",
+    "//packages/ecma402-abstract/DisplayNames",
     "//:node_modules/@formatjs/intl-localematcher",
 ]
 
@@ -726,6 +728,7 @@ ts_run_binary(
         "//:node_modules/json-stable-stringify",
         "//:node_modules/minimist",
         "//packages/ecma402-abstract",
+        "//packages/ecma402-abstract/types",
     ],
     outs = CLDR_RAW_FILES,
     args = [
