+++ packages/intl-durationformat/BUILD.bazel
@@ -21,6 +21,7 @@ SRCS = glob([
 
 SRC_DEPS = [
     "//packages/ecma402-abstract",
+    "//packages/ecma402-abstract/types",
     "//:node_modules/@formatjs/intl-localematcher",
 ]
 
@@ -137,6 +138,7 @@ generate_src_file(
         "//:node_modules/json-stable-stringify",
         "//:node_modules/lodash-es",
         "//packages/ecma402-abstract",
+        "//packages/ecma402-abstract/types",
     ],
     tags = ["cldr"],
     tool = "//packages/intl-durationformat/scripts:time-separators",
