+++ packages/intl-listformat/BUILD.bazel
@@ -38,6 +38,7 @@ TESTS = glob([
 
 SRC_DEPS = [
     "//packages/ecma402-abstract",
+    "//packages/ecma402-abstract/types",
     "//:node_modules/@formatjs/intl-localematcher",
 ]
 
@@ -729,6 +730,7 @@ ts_run_binary(
         "//:node_modules/lodash-es",
         "//:node_modules/minimist",
         "//packages/ecma402-abstract",
+        "//packages/ecma402-abstract/types",
     ],
     outs = CLDR_RAW_FILES,
     args = [
