+++ packages/intl-relativetimeformat/BUILD.bazel
@@ -39,6 +39,7 @@ TESTS = glob([
 
 SRC_DEPS = [
     "//packages/ecma402-abstract",
+    "//packages/ecma402-abstract/types",
     "//:node_modules/@formatjs/intl-localematcher",
 ]
 
@@ -727,6 +728,7 @@ ts_run_binary(
         "//:node_modules/json-stable-stringify",
         "//:node_modules/minimist",
         "//packages/ecma402-abstract",
+        "//packages/ecma402-abstract/types",
     ],
     outs = CLDR_RAW_FILES,
     args = [
